refactor(sessions): split pi path resolution
